Add sapling prf_ock and generate_trapdoor
